move pec func
