+++ mitiq/pec/channels.py
@@ -15,6 +15,7 @@ from typing import List
 import numpy as np
 import numpy.typing as npt
 from cirq import CNOT, OP_TREE, Circuit, DensityMatrixSimulator, H, LineQubit
+from mitiq.utils import vector_to_matrix, matrix_to_vector, tensor_product
 
 
 def _max_ent_state_circuit(num_qubits: int) -> Circuit:
@@ -86,31 +87,6 @@ def _operation_to_choi(operation_tree: OP_TREE) -> npt.NDArray[np.complex64]:
     return _circuit_to_choi(circuit)
 
 
-def tensor_product(
-    *args: npt.NDArray[np.complex64],
-) -> npt.NDArray[np.complex64]:
-    """Returns the Kronecker product of the input array-like arguments.
-    This is a generalization of the binary function
-    ``numpy.kron(arg_a, arg_b)`` to the case of an arbitrary number of
-    arguments.
-    """
-    if args == ():
-        raise TypeError("tensor_product() requires at least one argument.")
-
-    val = args[0]
-    for term in args[1:]:
-        val = np.kron(val, term)
-    return val
-
-
-def matrix_to_vector(
-    density_matrix: npt.NDArray[np.complex64],
-) -> npt.NDArray[np.complex64]:
-    r"""Reshapes a :math:`d \times d` density matrix into a
-    :math:`d^2`-dimensional state vector, according to the rule:
-    :math:`|i \rangle\langle j| \rightarrow |i,j \rangle`.
-    """
-    return density_matrix.flatten()
 
 
 def _safe_sqrt(
@@ -124,20 +100,6 @@ def _safe_sqrt(
         raise ValueError(error_str)
     return square_root
 
-
-def vector_to_matrix(
-    vector: npt.NDArray[np.complex64],
-) -> npt.NDArray[np.complex64]:
-    r"""Reshapes a :math:`d^2`-dimensional state vector into a
-    :math:`d \times d` density matrix, according to the rule:
-    :math:`|i,j \rangle \rightarrow |i \rangle\langle j|`.
-    """
-    error_str = (
-        "The expected dimension of the input vector must be a"
-        f" square number but is {vector.size}."
-    )
-    dim = _safe_sqrt(vector.size, error_str)
-    return vector.reshape(dim, dim)
 
 
 def kraus_to_super(
